Replace auth_x11 module with auth_stack
