+++ Makefile.am
@@ -143,23 +143,6 @@ auth_x11_SOURCES = \
 auth_x11_CPPFLAGS = $(macros) $(FONTCONFIG_CFLAGS) $(XFT_CFLAGS) $(LIBBSD_CFLAGS)
 auth_x11_LDADD = $(FONTCONFIG_LIBS) $(XFT_LIBS) $(LIBBSD_LIBS)
 
-helpers_PROGRAMS += \
-	auth_stack
-auth_stack_SOURCES = \
-	env_info.c env_info.h \
-	env_settings.c env_settings.h \
-	helpers/authproto.c helpers/authproto.h \
-	helpers/auth_stack.c \
-	helpers/monitors.c helpers/monitors.h \
-	logging.c logging.h \
-	mlock_page.h \
-	util.c util.h \
-	wait_pgrp.c wait_pgrp.h \
-	wm_properties.c wm_properties.h \
-	xscreensaver_api.c xscreensaver_api.h
-auth_stack_CPPFLAGS = $(macros) $(FONTCONFIG_CFLAGS) $(XFT_CFLAGS) $(LIBBSD_CFLAGS)
-auth_stack_LDADD = $(FONTCONFIG_LIBS) $(XFT_LIBS) $(LIBBSD_LIBS)
-
 if HAVE_PAM
 helpers_PROGRAMS += \
 	authproto_pam
